Add api method to get a list of all connected blocks
This allows other mods to iterate over all blocks the controller is
connected to without worrying about configured channels. With this
they can basically piggy-back onto the existing XNet network.

The new getConnectedBlockPositions method is very similiar to the
findConnectedBlocks one, but I wanted to keep this pull request
small and avoid having to introduce ConnectedBlockClientInfo to
the API. Also SidedPos relays exactly the minimum amount of
information to the caller, which is probably preferrable to avoid
potentially unnecessary overhead on each call.